Move CORS functionalities into a seperate class
